feat(win): reuse dark mode code
