+++ winio-ui-windows-common/Cargo.toml
@@ -13,6 +13,7 @@ winio-primitive = { workspace = true }
 
 compio = { workspace = true }
 
+cfg-if = { workspace = true }
 image = { workspace = true, default-features = false }
 widestring = { workspace = true }
 
@@ -35,5 +36,11 @@ windows-sys = { workspace = true, features = [
 ] }
 windows-numerics = "0.2"
 
+slim-detours-sys = { version = "0.3", optional = true }
+sync-unsafe-cell = { version = "0.1", optional = true }
+
 [build-dependencies]
 embed-resource = "3"
+
+[features]
+dark-mode = ["dep:slim-detours-sys", "dep:sync-unsafe-cell"]
