+++ winio-ui-winui/Cargo.toml
@@ -48,3 +48,6 @@ windows = { workspace = true, features = [
     "Win32_Graphics_Direct3D11",
     "Win32_Graphics_Dxgi",
 ] }
+
+[features]
+win32-dark-mode = ["winio-ui-windows-common/dark-mode"]
